docs(tutorial): fold closure note into paragraph; relocate test next to tutorial pipeline tests
